Remove legacy lessons activity
