+++ features/lessons/src/main/kotlin/app/ss/lessons/LessonsScreenUi.kt
@@ -56,19 +56,15 @@ import androidx.compose.ui.viewinterop.AndroidView
 import androidx.core.content.ContextCompat
 import app.ss.design.compose.theme.SsTheme
 import app.ss.design.compose.widget.scaffold.SsScaffold
-import com.cryart.sabbathschool.lessons.ui.lessons.LessonsTopBar
-import com.cryart.sabbathschool.lessons.ui.lessons.MIN_SOLID_ALPHA
-import com.cryart.sabbathschool.lessons.ui.lessons.ScrollAlpha
-import com.cryart.sabbathschool.lessons.ui.lessons.components.LessonItemsSpec
-import com.cryart.sabbathschool.lessons.ui.lessons.components.LessonsFooterSpec
-import com.cryart.sabbathschool.lessons.ui.lessons.components.footer
-import com.cryart.sabbathschool.lessons.ui.lessons.components.footerBackgroundColor
-import com.cryart.sabbathschool.lessons.ui.lessons.components.lessons
-import com.cryart.sabbathschool.lessons.ui.lessons.components.loading
-import com.cryart.sabbathschool.lessons.ui.lessons.components.quarterlyInfo
-import com.cryart.sabbathschool.lessons.ui.lessons.components.spec.toSpec
-import com.cryart.sabbathschool.lessons.ui.lessons.components.toSpec
-import com.cryart.sabbathschool.lessons.ui.lessons.rememberScrollAlpha
+import app.ss.lessons.components.LessonItemsSpec
+import app.ss.lessons.components.LessonsFooterSpec
+import app.ss.lessons.components.footer
+import app.ss.lessons.components.footerBackgroundColor
+import app.ss.lessons.components.lessons
+import app.ss.lessons.components.loading
+import app.ss.lessons.components.quarterlyInfo
+import app.ss.lessons.components.spec.toSpec
+import app.ss.lessons.components.toSpec
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.overlay.LocalOverlayHost
 import dagger.hilt.components.SingletonComponent
