+++ features/lessons/src/main/kotlin/app/ss/lessons/components/LessonItem.kt
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.cryart.sabbathschool.lessons.ui.lessons.components
+package app.ss.lessons.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
